refactor dpm campaign assignment routes
